feat(monorepo): realse script
