+++ package.json
@@ -80,6 +80,7 @@
     "@types/compression": "^1.7.0",
     "@types/cors": "^2.8.7",
     "@types/minimatch": "^3.0.3",
+    "@types/semver": "^7.3.4",
     "@types/ws": "^7.2.6",
     "chalk": "^4.1.0",
     "compression": "^1.7.4",
@@ -98,7 +99,7 @@
     "prismjs": "1.21.0",
     "puppeteer": "^5.5.0",
     "rxjs": "^6.6.2",
-    "semver": "^7.3.2",
+    "semver": "^7.3.4",
     "ws": "^7.3.1",
     "zone.js": "^0.10.3"
   },
