+++ tools/package.json
@@ -1,15 +1,12 @@
 {
-  "name": "scullyio/tooling",
+  "name": "tools",
   "version": "1.0.0",
-  "license": "MIT",
-  "url": "",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/scullyio/scully.git"
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "engines": {
-    "": "This is for building scully itself, using is fine with 12+",
-    "node": ">=14"
-  },
-  "type": "module"
+  "type": "module",
+  "author": "",
+  "license": "ISC"
 }
