+++ tools/tsconfig.tools.json
@@ -1,10 +1,11 @@
 {
-  "extends": "../tsconfig.json",
   "compilerOptions": {
-    "outDir": "../dist/out-tsc/tools",
+    "outDir": "../dist/tools",
     "rootDir": ".",
-    "module": "commonjs",
-    "target": "es5",
+    "target": "es2020",
+    "moduleResolution": "Node",
+    "module": "CommonJS",
+    "lib": ["ES2020"],
     "types": ["node"]
   },
   "include": ["**/*.ts"]
